Issue #765: 問い合わせページをpage.tsxとtemplate/index.tsxに分ける
